# fern/apis/summarize/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Summarize API",
    "description": "Transcribe and summarize videos from any source using any OpenAI-compatible LLM.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/summarize": {
      "post": {
        "summary": "Summarize a source",
        "description": "Summarize a video from a URL or file path. All fields from the CLI are supported. The response includes the generated summary in the requested format (markdown, json, or html).",
        "operationId": "summarize_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/summarize/upload": {
      "post": {
        "summary": "Summarize an uploaded file",
        "description": "Accepts video files (.mp4, .mp3, .wav, .m4a, .webm) or text files (.txt, .md, .vtt, .srt, .csv, .log, .rst, .html, .xml, .json). Text files bypass audio processing entirely. Maximum upload size is 500 MB.",
        "operationId": "summarize_upload_summarize_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_summarize_upload_summarize_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeResponse"
                }
              }
            }
          },
          "413": {
            "description": "File too large"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/summarize/batch": {
      "post": {
        "summary": "Batch summarize",
        "description": "Summarize multiple sources in one request. Each source is processed sequentially. Results are returned in the same order as the input sources list.",
        "operationId": "summarize_batch_summarize_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SummarizeRequest": {
        "properties": {
          "source": {
            "description": "Video URL or file path",
            "title": "Source",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "enum": [
                  "YouTube Video",
                  "Video URL",
                  "Google Drive Video Link",
                  "Dropbox Video Link",
                  "Local File",
                  "TXT"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "YouTube Video",
            "description": "Source type",
            "title": "Type"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider name from config",
            "title": "Provider"
          },
          "prompt_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Summary style",
            "title": "Prompt Type"
          },
          "chunk_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Characters per chunk",
            "maximum": 500000,
            "minimum": 100,
            "title": "Chunk Size"
          },
          "parallel_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Concurrent API requests",
            "maximum": 200,
            "minimum": 1,
            "title": "Parallel Calls"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Max output tokens per chunk",
            "maximum": 1000000,
            "minimum": 1,
            "title": "Max Tokens"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Caption/transcription language",
            "title": "Language"
          },
          "output_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Summary output language",
            "title": "Output Language"
          },
          "force_download": {
            "default": false,
            "description": "Skip captions, download audio",
            "title": "Force Download",
            "type": "boolean"
          },
          "transcription": {
            "anyOf": [
              {
                "enum": [
                  "Cloud Whisper",
                  "Local Whisper"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cloud Whisper or Local Whisper",
            "title": "Transcription"
          },
          "whisper_model": {
            "anyOf": [
              {
                "enum": [
                  "tiny",
                  "base",
                  "small",
                  "medium",
                  "large"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whisper model size",
            "title": "Whisper Model"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Playback speed for audio preprocessing or visual-mode video",
            "exclusiveMinimum": 0,
            "maximum": 10,
            "title": "Speed"
          },
          "output_format": {
            "anyOf": [
              {
                "enum": [
                  "markdown",
                  "json",
                  "html"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "markdown",
            "description": "markdown, json, or html",
            "title": "Output Format"
          },
          "visual": {
            "default": false,
            "description": "Send video directly to vision model",
            "title": "Visual",
            "type": "boolean"
          },
          "use_proxy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Route through the configured HTTP proxy",
            "title": "Use Proxy"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override API key",
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override API base URL",
            "title": "Base Url"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override model name",
            "title": "Model"
          },
          "cobalt_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cobalt base URL",
            "title": "Cobalt Url"
          },
          "verbose": {
            "default": false,
            "description": "Verbose progress output",
            "title": "Verbose",
            "type": "boolean"
          }
        },
        "required": [
          "source"
        ],
        "title": "SummarizeRequest",
        "type": "object"
      },
      "SummarizeResponse": {
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Type"
          },
          "processing_time_seconds": {
            "title": "Processing Time Seconds",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          }
        },
        "required": [
          "success",
          "source",
          "summary",
          "format",
          "processing_time_seconds"
        ],
        "title": "SummarizeResponse",
        "type": "object"
      },
...
    }
  }
}